@@ -1,9 +1,8 @@
 {
-  "name": "www",
+  "name": "my-project",
   "version": "1.0.0",
-  "type": "module",
   "private": true,
-  "description": "",
+  "type": "module",
   "scripts": {
     "dev": "vite --host",
     "build": "vite build",
@@ -12,24 +11,21 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "pinia": "^2.1.7",
-    "vue": "^3.4.31",
-    "vue-router": "^4.4.0",
-    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.2/xlsx-0.20.2.tgz"
+    "flowbite": "^2.5.1",
+    "pinia": "^2.2.2",
+    "vue": "^3.4.38",
+    "vue-router": "^4.4.3"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.10.3",
+    "@rushstack/eslint-patch": "^1.8.0",
     "@vitejs/plugin-vue": "^5.0.5",
-    "@vitejs/plugin-vue-jsx": "^4.0.0",
     "@vue/eslint-config-prettier": "^9.0.0",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^9.6.0",
-    "eslint-plugin-vue": "^9.27.0",
-    "postcss": "^8.4.39",
-    "prettier": "^3.3.2",
-    "tailwindcss": "^3.4.4",
-    "vite": "^5.3.3"
-  },
-  "author": "",
-  "license": "ISC"
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.9.1",
+    "eslint-plugin-vue": "^9.23.0",
+    "postcss": "^8.4.41",
+    "prettier": "^3.3.3",
+    "tailwindcss": "^3.4.10",
+    "vite": "^5.4.2"
+  }
 }
