@@ -1,17 +1,15 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
     "./index.html",
     "./src/**/*.{vue,js,ts,jsx,tsx}",
+    "./node_modules/flowbite/**/*.js",
   ],
   theme: {
-    extend: {
-      fontFamily: {
-        "playfair": ['Playfair Display', 'serif'],
-        "lato": ['Lato', 'sans-serif']
-      }
-    },
+    extend: {},
   },
-  plugins: [],
+  plugins: [
+    require('flowbite/plugin')
+  ],
 }
 
