@@ -1,10 +1,12 @@
-import './index.css'
+import './assets/main.css'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
 import App from './App.vue'
 import router from './router'
+
+import './index.css'
 
 const app = createApp(App)
 
